dom/virtualfilesystem/test/mochitest/test_virtualfilesystem_mount.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1026350
-->
<head>
<meta charset="utf-8">
<title>Test Inputport Connection Event</title>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1187723">Test FileSystemProvider Event
</a>
<script type="application/javascript;version=1.8">

"use strict";

SimpleTest.waitForExplicitFinish();

function compareMountOptionAndFileSystemInfo(option, info) {
  for (let name of ["fileSystemId", "displayName", "writable", "openedFilesLimit"]) {
    is(option[name], info[name], name + " should be the same");
  }
}

var mountOption = {fileSystemId: "dummyId",
                   displayName: "dummyFileSystem",
                   writable: false,
                   openedFilesLimit: 5};
var mountOption1 = {fileSystemId: "dummyId1",
                   displayName: "dummyFileSystem1",
                   writable: true,
                   openedFilesLimit: 10};

function testMount() {
  return new Promise(function(resolve, reject) {
    navigator.fileSystemProvider.mount(mountOption).then(
      function() {
        ok(true, "mount ok.");
        var fileSystemInfo = navigator.fileSystemProvider.get(mountOption.fileSystemId);
        ok(fileSystemInfo, "got file system info");
        compareMountOptionAndFileSystemInfo(fileSystemInfo, mountOption);

        navigator.fileSystemProvider.mount(mountOption1).then(
          function() {
            var fileSystemInfos = navigator.fileSystemProvider.getAll();
            ok(fileSystemInfos, "got all file systems");
            compareMountOptionAndFileSystemInfo(fileSystemInfos[0], mountOption);
            compareMountOptionAndFileSystemInfo(fileSystemInfos[1], mountOption1);
            navigator.fileSystemProvider.unmount({fileSystemId:mountOption1.fileSystemId}).then(
              function() {
                navigator.fileSystemProvider.unmount({fileSystemId:mountOption.fileSystemId}).then(
                  function() {
                    ok(true, "unmount all file systems ok")
                    resolve();
                  }
                );
              }
            );
          },
          function(aError) {
            ok(false, "Fail to mount: " + aError);
            resolve();
          }
        );
      },
      function(aError) {
        ok(false, "Fail to mount: " + aError);
        resolve();
      }
    );
  });
}


function runTest() {
  ok(navigator.fileSystemProvider, "should have navigator.fileSystemProvider");

  testMount()
  .then(function() {
    info("test finished");
    SimpleTest.finish();
  });
}

SpecialPowers.pushPrefEnv({"set": [["device.filesystemprovider.enabled", true],
                                    //to ignore app scope check.
                                   ["dom.ignore_webidl_scope_checks", true]]}, function() {
  SpecialPowers.pushPermissions(
    [{"type": "filesystemprovider", "allow": true, "context": document}], runTest);
});

 </script>
 </pre>
 </body>
 </html>
